requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
pydantic==2.9.2
httpx==0.27.2
loguru==0.7.2
mss==9.0.2
Pillow==10.4.0
pyautogui==0.9.54
